requirements.txt
fastapi==0.110.0
uvicorn==0.15.0
requests==2.22.0
beautifulsoup4==4.7.1
pydantic==2.6.3
aiohttp==3.8.1
gql==3.0.0
requests_toolbelt==1.0.0
google-cloud-storage==1.33.0
regex==2023.10.3
joblib==1.2.0
pytz
